Add UI to set task ownership in edit_todo.
Note: this changelist breaks "mass-edit" in that if a user selects
several tasks, edits them and presses save, Ocsigen will give out a
parameter error.  The fix for this will require a big refactoring of
the todo editor.  This refactoring will improve the code
maintainability though and the UI will be more intuitive.